+++ src/pages/user/Terms.js
@@ -1,23 +1,14 @@
-import React, { useEffect, useState } from 'react';
-import { RootUrl } from '../../api/RootUrl'; // ../../api 추가
-import { TERMS_PATH } from '../../requestPath'; // ../../ 추가
-import { selectTerms } from '../../api/UserApi'; // ../../ 추가
+// src/pages/user/ChekPass.js
+import React from 'react';
+import Terms from '../../components/user/Terms';
 
-const Terms = () => {
-    const [terms, setTerms] = useState({ terms: '', privacy: '', age: '' });
-
-    useEffect(() => {
-        selectTerms(TERMS_PATH, setTerms);
-    }, []);
-
-    return (
-        <div>
-            <h2>약관</h2>
-            <div>{terms.terms}</div>
-            <div>{terms.privacy}</div>
-            <div>{terms.age}</div>
-        </div>
-    );
+const terms = () => {
+  return (
+    <div>
+      <h1>Terms</h1>
+      <Terms />
+    </div>
+  );
 };
 
-export default Terms;
+export default terms;
