+++ src/styles/Terms.scss
@@ -1,0 +1,81 @@
+// src/styles/Terms.scss
+
+.Terms {
+    max-width: 400px;
+    margin: 0 auto;
+    padding: 2rem;
+    border: 1px solid #e5e5e5;
+    border-radius: 8px;
+    background-color: #fff;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    font-family: Arial, sans-serif;
+  
+    h2 {
+      margin-bottom: 1rem;
+      font-size: 1.5rem;
+      font-weight: bold;
+      text-align: center;
+    }
+  
+    .checkbox-group {
+      margin-bottom: 1rem;
+  
+      label {
+        display: flex;
+        align-items: center;
+        cursor: pointer;
+        margin-bottom: 0.5rem;
+  
+        input {
+          margin-right: 0.5rem;
+        }
+  
+        &.optional {
+          color: #777;
+        }
+      }
+  
+      .content {
+        display: none;
+        padding: 0.5rem;
+        border: 1px solid #ddd;
+        border-radius: 4px;
+        margin-top: 0.5rem;
+        background-color: #f9f9f9;
+  
+        &.expanded {
+          display: block;
+        }
+      }
+    }
+  
+    .checkbox-all {
+      margin-bottom: 1rem;
+      border-bottom: 1px solid #e5e5e5;
+      padding-bottom: 1rem;
+      display: flex;
+      align-items: center;
+  
+      input {
+        margin-right: 0.5rem;
+      }
+    }
+  
+    .submit-btn {
+      display: block;
+      width: 100%;
+      padding: 0.75rem;
+      background-color: #000;
+      color: #fff;
+      text-align: center;
+      border: none;
+      border-radius: 4px;
+      cursor: pointer;
+      font-size: 1rem;
+  
+      &:hover {
+        background-color: #333;
+      }
+    }
+  }
+
